+++ example/lambdacd_git/example/simple_pipeline.clj
@@ -36,7 +36,6 @@
   (let [home-dir (io/file "/tmp/foo")
         config   {:home-dir      home-dir}
         pipeline (lambdacd/assemble-pipeline pipeline-structure config)]
-    (core/init-ssh!)
     (runners/start-one-run-after-another pipeline)
     (ring-server/serve (routes
                          (ui/ui-for pipeline)