+++ test/lambdacd_git/end_to_end_test.clj
@@ -56,7 +56,6 @@
 
 (deftest example-pipeline-test
   (testing "the example-pipeline"
-    (core/init-ssh!)
     (let [config                 {:home-dir (lambdacd-util/create-temp-dir)}
           pipeline               (lambdacd/assemble-pipeline simple-pipeline/pipeline-structure config)
           future-pipeline-result (future
@@ -67,7 +66,6 @@
 
 (deftest ^:e2e-with-auth end-to-end-with-auth-test
   (testing "a complete pipeline with all features against private repositories using https and ssh"
-    (core/init-ssh!)
     (doseq [repo-config [{:repo-uri (or (System/getenv "LAMBDACD_GIT_TESTREPO_SSH") "git@gitlab.com:flosell-test/testrepo.git")}
                          {:repo-uri (or (System/getenv "LAMBDACD_GIT_TESTREPO_HTTPS") "https://gitlab.com/flosell-test/testrepo.git")
                           :git      {:credentials-provider (UsernamePasswordCredentialsProvider. (System/getenv "LAMBDACD_GIT_TESTREPO_USERNAME")